Extract the large GraphitiLLMConfig class to its own dedicated file
As a developer, I need the LLM configuration logic separated into its own file since it's currently 317 lines and represents a major component with single responsibility (LLM client configuration).
