+++ mcp_server/src/config/__init__.py
@@ -1,0 +1,7 @@
+"""Configuration modules for the Graphiti MCP server."""
+
+from .llm_config import GraphitiLLMConfig
+
+__all__ = [
+    "GraphitiLLMConfig",
+]
